immersive-commerce-engine/3d-asset-management-service/requirements.txt
django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.1.1
